+++ spine-ios/Example/Spine iOS Example/AnimationStateEvents.swift
@@ -1,10 +1,3 @@
-//
-//  AnimationStateEvents.swift
-//  Spine iOS Example
-//
-//  Created by Denis Andrašec on 15.05.24.
-//
-
 import SwiftUI
 import Spine
 import SpineCppLite