+++ spine-ios/Example/Spine iOS Example/MainView.swift
@@ -1,35 +1,50 @@
-//
-//  MainView.swift
-//  Spine iOS Example
-//
-//  Created by Denis Andrašec on 08.05.24.
-//
-
 import SwiftUI
+import Spine
 
 struct MainView: View {
     var body: some View {
         List {
-            NavigationLink("SimpleAnimation") {
-                SimpleAnimation()
+            Section {
+                NavigationLink("Simple Animation") {
+                    SimpleAnimation()
+                }
+                NavigationLink("Play/Pause") {
+                    PlayPauseAnimation()
+                }
+                NavigationLink("Animation State Listener") {
+                    AnimationStateEvents()
+                }
+                NavigationLink("Debug Rendering") {
+                    DebugRendering()
+                }
+                NavigationLink("Dress Up") {
+                    DressUp()
+                }
+                NavigationLink("IK Following") {
+                    IKFollowing()
+                }
+                NavigationLink("Physics") {
+                    Physics()
+                }
+            } header: {
+                Text("Swift + SwiftUI")
             }
-            NavigationLink("Play/Pause") {
-                PlayPauseAnimation()
-            }
-            NavigationLink("Animation State Listener") {
-                AnimationStateEvents()
-            }
-            NavigationLink("Debug Rendering") {
-                DebugRendering()
-            }
-            NavigationLink("Dress Up") {
-                DressUp()
-            }
-            NavigationLink("IK Following") {
-                IKFollowing()
-            }
-            NavigationLink("Physics") {
-                Physics()
+            Section {
+                NavigationLink("Simple Animation") {
+                    SimpleAnimationViewControllerRepresentable()
+                        .navigationTitle("Simple Animation")
+                        .navigationBarTitleDisplayMode(.inline)
+                }
+            } header: {
+                Text("ObjC + UIKit")
+            } footer: {
+                HStack {
+                    Spacer()
+                    Text("Spine \(Spine.version)")
+                        .font(.footnote)
+                        .foregroundStyle(.secondary)
+                    Spacer()
+                }
             }
         }
         .navigationTitle("Spine Examples")