+++ spine-ios/Example/Spine iOS Example/Physics.swift
@@ -1,10 +1,3 @@
-//
-//  Physics.swift
-//  Spine iOS Example
-//
-//  Created by Denis Andrašec on 22.05.24.
-//
-
 import SwiftUI
 import Spine