+++ spine-ios/Example/Spine iOS Example/PlayPauseAnimation.swift
@@ -1,10 +1,3 @@
-//
-//  PlayPause.swift
-//  Spine iOS Example
-//
-//  Created by Denis Andrašec on 08.05.24.
-//
-
 import SwiftUI
 import Spine