+++ spine-ios/Example/Spine iOS Example/SimpleAnimation.swift
@@ -1,10 +1,3 @@
-//
-//  ContentView.swift
-//  Spine iOS Example
-//
-//  Created by Denis Andrašec on 17.04.24.
-//
-
 import SwiftUI
 import Spine