+++ spine-ios/Example/Spine iOS Example/SimpleAnimationViewControllerRepresentable.swift
@@ -1,0 +1,13 @@
+import SwiftUI
+
+struct SimpleAnimationViewControllerRepresentable: UIViewControllerRepresentable {
+    typealias UIViewControllerType = SimpleAnimationViewController
+    
+    func makeUIViewController(context: Context) -> SimpleAnimationViewController {
+        return SimpleAnimationViewController()
+    }
+    
+    func updateUIViewController(_ uiViewController: SimpleAnimationViewController, context: Context) {
+        //
+    }
+}