+++ spine-ios/CHANGELOG.md
@@ -15,3 +15,4 @@
 - Add `DebugRendering` sample (#11)
 - Move remaining files to SPM package (#12)
 - Load assets from different sources (#13)
+- Add `Obj-C` + `UIKit` sample (#14)